pipeline/requirements.txt
requests
pandas
psycopg2-binary
python-dotenv
pandarallel
openai
boto3